Rename `Gw2AccountSubscriptionProvider.tsx` to `use-gw2-subscription.ts`
